Delete EventService
