+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,8 +1,12 @@
 package ru.yandex.practicum.filmorate.storage;
 
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
 
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Map;
 
 public interface FilmStorage extends BaseStorage<Film> {
 
@@ -16,5 +20,13 @@ public interface FilmStorage extends BaseStorage<Film> {
     List<Film> findFilms(String query, String by);
 
     List<Film> getRecommendation(long id);
+
+    Map<Long, LinkedHashSet<Genre>> getFilmsGenres(List<Film> films);
+
+    Map<Long, LinkedHashSet<Genre>> getAllFilmGenres();
+
+    Map<Long, LinkedHashSet<Director>> getAllFilmDirectors();
+
+    Map<Long, LinkedHashSet<Director>> getFilmsDirectors(List<Film> films);
 }
